requirements.txt
biopython
matplotlib
scikit-bio
ete3